Split the agent-chat notification helper out of AppDelegate+AgentChat.swift
Two concurrent merges (#7216 + #7723) combined the file to 511 lines,
past the 500-line tracking threshold with no budget entry, failing the
guard on every PR's merge preview. Both files now sit under 500.
