+++ Sources/AppDelegate+AgentChat.swift
@@ -205,50 +205,6 @@ extension AppDelegate {
         return tabManager.tabs.first { !beforeIds.contains($0.id) } ?? tabManager.selectedWorkspace
     }
 
-    private func postAgentChatServerUnavailableNotification(
-        workspace: Workspace?,
-        agentChat: CmuxAgentChatConfiguration
-    ) {
-        let body: String
-        if let startCommand = agentChat.startCommand {
-            let format = String(
-                localized: "notification.agentChat.serverUnavailable.bodyWithCommand",
-                defaultValue: "cmux couldn't reach %@. Start it with: %@"
-            )
-            body = String(format: format, agentChat.url.absoluteString, startCommand)
-        } else {
-            let format = String(
-                localized: "notification.agentChat.serverUnavailable.bodyDefault",
-                defaultValue: "cmux couldn't reach %@. Start the server with cmux-chat or configure agentChat.startCommand in cmux.json."
-            )
-            body = String(format: format, agentChat.url.absoluteString)
-        }
-        // No workspace = owned launch failed; anchor to the focused workspace.
-        guard let anchorTabId = workspace?.id ?? activeTabManagerForCommands(preferredWindow: nil)?.selectedTabId else {
-            return
-        }
-        let subtitle = workspace == nil
-            ? String(
-                localized: "notification.agentChat.serverUnavailable.subtitleLaunchFailed",
-                defaultValue: "Could not start Agent Chat"
-            )
-            : String(
-                localized: "notification.agentChat.serverUnavailable.subtitle",
-                defaultValue: "Opened Agent Chat"
-            )
-        TerminalNotificationStore.shared.addNotification(
-            tabId: anchorTabId,
-            surfaceId: workspace?.focusedPanelId,
-            title: String(
-                localized: "notification.agentChat.serverUnavailable.title",
-                defaultValue: "Agent chat server isn't running"
-            ),
-            subtitle: subtitle,
-            body: body,
-            cooldownKey: "agent-chat-server-unavailable.\(agentChat.url.absoluteString)",
-            cooldownInterval: 30
-        )
-    }
 
     private func ensureAgentChatServerAvailable(
         _ agentChat: CmuxAgentChatConfiguration,
